Move tests that needs a FactoryBinder
Issue #207
Some tests need a FactoryBinder to work so I moved them into the proper
module
