+++ pcap4j-core/pom.xml
@@ -60,18 +60,6 @@
       <artifactId>powermock-api-mockito</artifactId>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>org.pcap4j</groupId>
-      <artifactId>pcap4j-packetfactory-static</artifactId>
-      <version>1.6.3</version>
-      <exclusions>
-        <exclusion>
-          <groupId>org.pcap4j</groupId>
-          <artifactId>pcap4j-core</artifactId>
-        </exclusion>
-      </exclusions>
-      <scope>test</scope>
-    </dependency>
   </dependencies>
 
   <build>
@@ -86,25 +74,6 @@
     </resources>
 
     <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-antrun-plugin</artifactId>
-        <executions>
-          <execution>
-            <phase>process-classes</phase>
-            <goals>
-              <goal>run</goal>
-            </goals>
-            <configuration>
-              <target>
-                <delete>
-                  <fileset dir="${project.build.outputDirectory}" includes="**/PacketFactoryBinder.class" />
-                </delete>
-              </target>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
       <!-- Provides a module name since Java 9 modularity -->
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
@@ -159,9 +128,6 @@
             <groupId>org.apache.maven.plugins</groupId>
             <artifactId>maven-source-plugin</artifactId>
             <configuration>
-              <excludes>
-                <exclude>**/PacketFactoryBinder.java</exclude>
-              </excludes>
               <includePom>false</includePom>
             </configuration>
             <executions>
